chore: move types and utils
