+++ ui/app/(mission-locale)/mission-locale/MissionLocaleClient.tsx
@@ -4,10 +4,10 @@ import { useQuery } from "@tanstack/react-query";
 
 import { MissionLocaleDisplay } from "@/app/_components/mission-locale/MissionLocaleDisplay";
 import { MLHeader } from "@/app/_components/mission-locale/MLHeader";
-import { MonthsData } from "@/app/_components/mission-locale/types";
 import { PageWithSidebarSkeleton } from "@/app/_components/suspense/LoadingSkeletons";
 import { SuspenseWrapper } from "@/app/_components/suspense/SuspenseWrapper";
 import { _get } from "@/common/httpClient";
+import { MonthsData } from "@/common/types/ruptures";
 
 export default function MissionLocaleClient() {
   const { data } = useQuery<MonthsData>(
