+++ ui/app/admin/mission-locale/[id]/edit/MissionLocaleAdminEditClient.tsx
@@ -6,10 +6,10 @@ import { useQuery } from "@tanstack/react-query";
 import { DsfrLink } from "@/app/_components/link/DsfrLink";
 import { MissionLocaleDisplay } from "@/app/_components/mission-locale/MissionLocaleDisplay";
 import { ModalAdminSyncBrevo } from "@/app/_components/mission-locale/modal/ModalAdminSyncBrevo";
-import { MonthsData } from "@/app/_components/mission-locale/types";
 import { PageWithSidebarSkeleton } from "@/app/_components/suspense/LoadingSkeletons";
 import { SuspenseWrapper } from "@/app/_components/suspense/SuspenseWrapper";
 import { _get } from "@/common/httpClient";
+import { MonthsData } from "@/common/types/ruptures";
 
 export default function MissionLocaleAdminEditClient({ id }: { id: string }) {
   const { data: missionLocaleData } = useQuery(
